Move editor options out of editorCommon
