+++ src/vs/editor/browser/viewParts/overviewRuler/overviewRulerImpl.ts
@@ -5,11 +5,12 @@
 'use strict';
 
 import { FastDomNode, createFastDomNode } from 'vs/base/browser/fastDomNode';
-import { OverviewRulerPosition, OverviewRulerLane, OverviewRulerZone, ColorZone, ThemeType } from 'vs/editor/common/editorCommon';
+import { OverviewRulerLane, OverviewRulerZone, ColorZone, ThemeType } from 'vs/editor/common/editorCommon';
 import { IDisposable } from 'vs/base/common/lifecycle';
 import * as browser from 'vs/base/browser/browser';
 import { OverviewZoneManager } from 'vs/editor/common/view/overviewZoneManager';
 import { Color } from 'vs/base/common/color';
+import { OverviewRulerPosition } from "vs/editor/common/config/editorOptions";
 
 export class OverviewRulerImpl {
