+++ src/vs/editor/browser/viewParts/viewCursors/viewCursors.ts
@@ -6,7 +6,6 @@
 'use strict';
 
 import 'vs/css!./viewCursors';
-import * as editorCommon from 'vs/editor/common/editorCommon';
 import { ClassNames } from 'vs/editor/browser/editorBrowser';
 import { ViewPart } from 'vs/editor/browser/view/viewPart';
 import { Position } from 'vs/editor/common/core/position';
@@ -18,14 +17,15 @@ import { TimeoutTimer, IntervalTimer } from 'vs/base/common/async';
 import * as viewEvents from 'vs/editor/common/view/viewEvents';
 import { registerThemingParticipant } from 'vs/platform/theme/common/themeService';
 import { editorCursor } from 'vs/editor/common/view/editorColorRegistry';
+import { TextEditorCursorBlinkingStyle, TextEditorCursorStyle } from "vs/editor/common/config/editorOptions";
 
 export class ViewCursors extends ViewPart {
 
 	static BLINK_INTERVAL = 500;
 
 	private _readOnly: boolean;
-	private _cursorBlinking: editorCommon.TextEditorCursorBlinkingStyle;
-	private _cursorStyle: editorCommon.TextEditorCursorStyle;
+	private _cursorBlinking: TextEditorCursorBlinkingStyle;
+	private _cursorStyle: TextEditorCursorStyle;
 	private _selectionIsEmpty: boolean;
 
 	private _isVisible: boolean;
@@ -198,12 +198,12 @@ export class ViewCursors extends ViewPart {
 
 	// ---- blinking logic
 
-	private _getCursorBlinking(): editorCommon.TextEditorCursorBlinkingStyle {
+	private _getCursorBlinking(): TextEditorCursorBlinkingStyle {
 		if (!this._editorHasFocus) {
-			return editorCommon.TextEditorCursorBlinkingStyle.Hidden;
+			return TextEditorCursorBlinkingStyle.Hidden;
 		}
 		if (this._readOnly || !this._primaryCursor.getIsInEditableRange()) {
-			return editorCommon.TextEditorCursorBlinkingStyle.Solid;
+			return TextEditorCursorBlinkingStyle.Solid;
 		}
 		return this._cursorBlinking;
 	}
@@ -215,8 +215,8 @@ export class ViewCursors extends ViewPart {
 		let blinkingStyle = this._getCursorBlinking();
 
 		// hidden and solid are special as they involve no animations
-		let isHidden = (blinkingStyle === editorCommon.TextEditorCursorBlinkingStyle.Hidden);
-		let isSolid = (blinkingStyle === editorCommon.TextEditorCursorBlinkingStyle.Solid);
+		let isHidden = (blinkingStyle === TextEditorCursorBlinkingStyle.Hidden);
+		let isSolid = (blinkingStyle === TextEditorCursorBlinkingStyle.Solid);
 
 		if (isHidden) {
 			this._hide();
@@ -228,7 +228,7 @@ export class ViewCursors extends ViewPart {
 		this._updateDomClassName();
 
 		if (!isHidden && !isSolid) {
-			if (blinkingStyle === editorCommon.TextEditorCursorBlinkingStyle.Blink) {
+			if (blinkingStyle === TextEditorCursorBlinkingStyle.Blink) {
 				// flat blinking is handled by JavaScript to save battery life due to Chromium step timing issue https://bugs.chromium.org/p/chromium/issues/detail?id=361587
 				this._cursorFlatBlinkInterval.cancelAndSet(() => {
 					if (this._isVisible) {
@@ -257,22 +257,22 @@ export class ViewCursors extends ViewPart {
 			result += ' has-selection';
 		}
 		switch (this._cursorStyle) {
-			case editorCommon.TextEditorCursorStyle.Line:
+			case TextEditorCursorStyle.Line:
 				result += ' cursor-line-style';
 				break;
-			case editorCommon.TextEditorCursorStyle.Block:
+			case TextEditorCursorStyle.Block:
 				result += ' cursor-block-style';
 				break;
-			case editorCommon.TextEditorCursorStyle.Underline:
+			case TextEditorCursorStyle.Underline:
 				result += ' cursor-underline-style';
 				break;
-			case editorCommon.TextEditorCursorStyle.LineThin:
+			case TextEditorCursorStyle.LineThin:
 				result += ' cursor-line-thin-style';
 				break;
-			case editorCommon.TextEditorCursorStyle.BlockOutline:
+			case TextEditorCursorStyle.BlockOutline:
 				result += ' cursor-block-outline-style';
 				break;
-			case editorCommon.TextEditorCursorStyle.UnderlineThin:
+			case TextEditorCursorStyle.UnderlineThin:
 				result += ' cursor-underline-thin-style';
 				break;
 			default:
@@ -280,19 +280,19 @@ export class ViewCursors extends ViewPart {
 		}
 		if (this._blinkingEnabled) {
 			switch (this._getCursorBlinking()) {
-				case editorCommon.TextEditorCursorBlinkingStyle.Blink:
+				case TextEditorCursorBlinkingStyle.Blink:
 					result += ' cursor-blink';
 					break;
-				case editorCommon.TextEditorCursorBlinkingStyle.Smooth:
+				case TextEditorCursorBlinkingStyle.Smooth:
 					result += ' cursor-smooth';
 					break;
-				case editorCommon.TextEditorCursorBlinkingStyle.Phase:
+				case TextEditorCursorBlinkingStyle.Phase:
 					result += ' cursor-phase';
 					break;
-				case editorCommon.TextEditorCursorBlinkingStyle.Expand:
+				case TextEditorCursorBlinkingStyle.Expand:
 					result += ' cursor-expand';
 					break;
-				case editorCommon.TextEditorCursorBlinkingStyle.Solid:
+				case TextEditorCursorBlinkingStyle.Solid:
 					result += ' cursor-solid';
 					break;
 				default:
