+++ src/vs/editor/browser/widget/embeddedCodeEditorWidget.ts
@@ -8,10 +8,11 @@ import * as objects from 'vs/base/common/objects';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { ICommandService } from 'vs/platform/commands/common/commands';
 import { IContextKeyService } from 'vs/platform/contextkey/common/contextkey';
-import { ICodeEditorWidgetCreationOptions, IConfigurationChangedEvent, IEditorOptions } from 'vs/editor/common/editorCommon';
+import { ICodeEditorWidgetCreationOptions } from 'vs/editor/common/editorCommon';
 import { ICodeEditorService } from 'vs/editor/common/services/codeEditorService';
 import { ICodeEditor } from 'vs/editor/browser/editorBrowser';
 import { CodeEditor } from 'vs/editor/browser/codeEditor';
+import { IConfigurationChangedEvent, IEditorOptions } from "vs/editor/common/config/editorOptions";
 
 export class EmbeddedCodeEditorWidget extends CodeEditor {
