+++ src/vs/editor/common/config/defaultConfig.ts
@@ -5,9 +5,9 @@
 'use strict';
 
 import * as nls from 'vs/nls';
-import { IEditorOptions } from 'vs/editor/common/editorCommon';
 import * as platform from 'vs/base/common/platform';
 import { USUAL_WORD_SEPARATORS } from 'vs/editor/common/model/wordHelper';
+import { IEditorOptions } from "vs/editor/common/config/editorOptions";
 
 export interface IConfiguration {
 	editor: IEditorOptions;
