+++ src/vs/editor/common/viewModel/characterHardWrappingLineMapper.ts
@@ -5,12 +5,12 @@
 'use strict';
 
 import * as strings from 'vs/base/common/strings';
-import { WrappingIndent } from 'vs/editor/common/editorCommon';
 import { PrefixSumComputer } from 'vs/editor/common/viewModel/prefixSumComputer';
 import { ILineMapperFactory, ILineMapping, OutputPosition } from 'vs/editor/common/viewModel/splitLinesCollection';
 import { CharCode } from 'vs/base/common/charCode';
 import { CharacterClassifier } from 'vs/editor/common/core/characterClassifier';
 import { toUint32Array } from 'vs/editor/common/core/uint';
+import { WrappingIndent } from "vs/editor/common/config/editorOptions";
 
 const enum CharacterClass {
 	NONE = 0,
