+++ src/vs/editor/contrib/hover/browser/hoverWidgets.ts
@@ -8,11 +8,12 @@ import { KeyCode } from 'vs/base/common/keyCodes';
 import { IKeyboardEvent } from 'vs/base/browser/keyboardEvent';
 import { toggleClass } from 'vs/base/browser/dom';
 import { Position } from 'vs/editor/common/core/position';
-import { IPosition, IConfigurationChangedEvent } from 'vs/editor/common/editorCommon';
+import { IPosition } from 'vs/editor/common/editorCommon';
 import * as editorBrowser from 'vs/editor/browser/editorBrowser';
 import { Widget } from 'vs/base/browser/ui/widget';
 import { DomScrollableElement } from 'vs/base/browser/ui/scrollbar/scrollableElement';
 import { IDisposable, dispose } from 'vs/base/common/lifecycle';
+import { IConfigurationChangedEvent } from "vs/editor/common/config/editorOptions";
 
 export class ContentHoverWidget extends Widget implements editorBrowser.IContentWidget {
