+++ src/vs/editor/contrib/selectionClipboard/electron-browser/selectionClipboard.ts
@@ -9,11 +9,12 @@ import { clipboard } from 'electron';
 import * as platform from 'vs/base/common/platform';
 import { ICodeEditor, IEditorMouseEvent } from 'vs/editor/browser/editorBrowser';
 import { Disposable } from 'vs/base/common/lifecycle';
-import { EndOfLinePreference, IEditorContribution, ICursorSelectionChangedEvent, IConfigurationChangedEvent } from 'vs/editor/common/editorCommon';
+import { EndOfLinePreference, IEditorContribution, ICursorSelectionChangedEvent } from 'vs/editor/common/editorCommon';
 import { editorContribution } from 'vs/editor/browser/editorBrowserExtensions';
 import { IContextKeyService } from 'vs/platform/contextkey/common/contextkey';
 import { RunOnceScheduler } from 'vs/base/common/async';
 import { Range } from 'vs/editor/common/core/range';
+import { IConfigurationChangedEvent } from "vs/editor/common/config/editorOptions";
 
 @editorContribution
 export class SelectionClipboard extends Disposable implements IEditorContribution {
