+++ src/vs/workbench/test/electron-browser/api/extHostTextEditor.test.ts
@@ -7,7 +7,7 @@
 import * as assert from 'assert';
 import { TPromise } from 'vs/base/common/winjs.base';
 import { TextEditorLineNumbersStyle } from 'vs/workbench/api/node/extHostTypes';
-import { TextEditorCursorStyle } from 'vs/editor/common/editorCommon';
+import { TextEditorCursorStyle } from 'vs/editor/common/config/editorOptions';
 import { IResolvedTextEditorConfiguration, ITextEditorConfigurationUpdate } from 'vs/workbench/api/node/mainThreadEditor';
 import { MainThreadEditorsShape } from 'vs/workbench/api/node/extHost.protocol';
 import { ExtHostTextEditorOptions, ExtHostTextEditor } from 'vs/workbench/api/node/extHostTextEditor';
